Auto merged
2004/03/09 14:38:49+00:00 cantab.net!aia21
Fix all occurences of printf with %ll length modifiers but 64 bit
arguments to typecast the arguments to (unsigned) long long to avoid
the warnings when compiling on 64 bit architectures.

(Logical change 1.306)

+++ libntfs/bitmap.c
@@ -165,7 +165,7 @@ static int ntfs_bitmap_set_bits_in_run(ntfs_attr *na, s64 start_bit, s64 count,
 			fprintf(stderr, "%s(): Eeek! Last buffer but count is "
 					"not zero (= %lli). Leaving "
 					"inconsistent metadata.\n",
-					__FUNCTION__, count);
+					__FUNCTION__, (long long)count);
 			err = EIO;
 			goto free_err_out;
 		}
